% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastlm.R
\name{get_slope}
\alias{get_intercept}
\alias{get_rsq}
\alias{get_slope}
\title{Retrieve Slope, Intercept, and R Squared}
\usage{
get_slope(x)
}
\arguments{
\item{x}{fastlm object}
}
\value{
numeric(1L)
}
\description{
Retrieve Slope, Intercept, and R Squared
}

